utils refactoring
