--- src/components/Select/utils/index.ts
@@ -1,3 +1,0 @@
-export * from './constants';
-export * from './listbox';
-export * from './types';
